charms/tensorboard-controller/requirements-unit.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --resolver=backtracking ./requirements-unit.in
#
attrs==22.1.0
    # via pytest
coverage==6.5.0
    # via -r ./requirements-unit.in
exceptiongroup==1.0.4
    # via pytest
iniconfig==1.1.1
    # via pytest
oci-image==1.0.0
    # via -r ./requirements.txt
ops==1.5.4
    # via -r ./requirements.txt
packaging==22.0
    # via pytest
pluggy==1.0.0
    # via pytest
pytest==7.2.0
    # via
    #   -r ./requirements-unit.in
    #   pytest-lazy-fixture
    #   pytest-mock
pytest-lazy-fixture==0.6.3
    # via -r ./requirements-unit.in
pytest-mock==3.10.0
    # via -r ./requirements-unit.in
pyyaml==6.0
    # via
    #   -r ./requirements.txt
    #   ops
tomli==2.0.1
    # via pytest
